+++ cmsafvis/NEWS.md
@@ -1,11 +1,19 @@
 # Changelog for cmsafvis package
 
+## 1.1.10
+
+- new config parameter for quicklook: 'unit' - define variable unit
+  - e.g., Percent or '%'
+- new color scales 'albedo' and 'albedo2' added 
+
 ## 1.1.9
 
 - plotKML package is back on CRAN und export of KML is possible again
 - fieldmean plot includes now all available years and not just years included
   in the climatology period
 - new color scale 'larry' added
+- get_basename_vis added to helper functions to avoid clash with   
+  cmsafops::get_basename
 
 ## 1.1.8